implement level fusion, low-bit integer tree, and zero-allocation execution
Summary of changes:
   - Implement bottom-up level fusion for early tree stages (levels 0..F):
     disjoint sub-trees of 2^F channels are evaluated vertically on-chip
     (CPU L1/L2 cache, GPU shared memory), eliminating DRAM round-trips for
     the most memory-intensive levels (~45-50% of total memory traffic).
   - Transition to global breadth-first ping-pong merges above level F.
   - CUDA backend partitions into time tiles with halo recomputation in
     shared memory, delivering bit-exact parity with the CPU.
   - Support packed low-bit inputs (1, 2, 4, 8, 16 bits) directly.
   - Store intermediate tree levels with bounded exact sums as uint8/uint16
     instead of float32, reducing memory traffic by up to 4x.
   - Remove intra-level cache tiling (kTiled, ScheduleType) after empirical
     benchmarks showed the plain coordinate loop already achieves 93-99% of
     raw copy bandwidth with near-optimal row reuse (1.14 reads/output).
   - Remove non-temporal streaming stores (simd_intrinsics.hpp), which degraded
     single-thread performance on x86 and do not apply to fused levels.
   - Pre-size and allocate all working memory (fused ping-pong scratch,
     unpack row, and running boxcar rows) at construction in a single
     std::vector<std::byte> workspace; execute() and stepper perform zero allocations.
   - Pin OpenMP parallel loops to instance thread count (num_threads(m_ws.nthreads))
     to guarantee thread safety against global OpenMP changes.
   - Eliminate FDMTExecConfig; pass fuse_levels and int_tree directly into
     FDMTCPU and FDMTCUDA constructors with optimal defaults.
   - Fix buffer comparison in CUDA parity tests to validate leading get_dmt_size()
     transform elements rather than total ping-pong buffer extent.
   - Add new tests: fdmt_alloc_cpu_t (asserts 0 allocations during execution) and
     fdmt_fused_tile_cpu_t (host emulation of CUDA tile geometry).
   - Add Performance Tuning guide (docs/pipeline_guide/performance.md) documenting
     the hybrid traversal architecture and benchmark results.

+++ lib/fdmt_fft_cuda.cu
@@ -4,6 +4,7 @@
 
 #include <algorithm>
 #include <cstddef>
+#include <cstdint>
 #include <format>
 #include <memory>
 #include <stdexcept>
@@ -58,7 +59,9 @@ __global__ void kernel_fill_window(const float* __restrict__ wf,
                                    int nbeams) {
     const auto t = static_cast<int>((blockIdx.x * blockDim.x) + threadIdx.x);
     const auto i_chan = static_cast<int>(blockIdx.y);
-    const auto i_beam = static_cast<int>(blockIdx.z);
+    // 64-bit: beam-strided offsets (i_beam * per-beam size) exceed
+    // int32 for large blocks with several beams.
+    const auto i_beam = static_cast<int64_t>(blockIdx.z);
     if (t >= n_fft || i_chan >= nchans || i_beam >= nbeams) {
         return;
     }
@@ -94,7 +97,9 @@ __global__ void kernel_update_overlap(const float* __restrict__ wf,
                                       int nbeams) {
     const auto i = static_cast<int>((blockIdx.x * blockDim.x) + threadIdx.x);
     const auto i_chan = static_cast<int>(blockIdx.y);
-    const auto i_beam = static_cast<int>(blockIdx.z);
+    // 64-bit: beam-strided offsets (i_beam * per-beam size) exceed
+    // int32 for large blocks with several beams.
+    const auto i_beam = static_cast<int64_t>(blockIdx.z);
     if (i >= overlap_len || i_chan >= nchans || i_beam >= nbeams) {
         return;
     }
@@ -122,8 +127,10 @@ kernel_init_fdmt_fft(const Cplx* __restrict__ spectra,
                      int n_bins,
                      int max_coords) {
     const auto k = static_cast<int>((blockIdx.x * blockDim.x) + threadIdx.x);
-    const auto i_sub  = static_cast<int>(blockIdx.y);
-    const auto i_beam = static_cast<int>(blockIdx.z);
+    const auto i_sub = static_cast<int>(blockIdx.y);
+    // 64-bit: beam-strided offsets (i_beam * per-beam size) exceed
+    // int32 for large blocks with several beams.
+    const auto i_beam = static_cast<int64_t>(blockIdx.z);
     if (k >= n_bins || i_sub >= nsubs) {
         return;
     }
@@ -152,7 +159,9 @@ __global__ void kernel_execute_iter_fft(const Cplx* __restrict__ state_in,
                                         int max_coords) {
     const auto linear =
         static_cast<int>((blockIdx.x * blockDim.x) + threadIdx.x);
-    const auto i_beam         = static_cast<int>(blockIdx.y);
+    // 64-bit: beam-strided offsets (i_beam * per-beam size) exceed
+    // int32 for large blocks with several beams.
+    const auto i_beam         = static_cast<int64_t>(blockIdx.y);
     const int max_iter_coords = (ncoords_sum_cur > ncoords_copy_cur)
                                     ? ncoords_sum_cur
                                     : ncoords_copy_cur;
@@ -206,7 +215,9 @@ __global__ void kernel_trim_scale(const float* __restrict__ time_out,
                                   float norm) {
     const auto t    = static_cast<int>((blockIdx.x * blockDim.x) + threadIdx.x);
     const auto i_dm = static_cast<int>(blockIdx.y);
-    const auto i_beam = static_cast<int>(blockIdx.z);
+    // 64-bit: beam-strided offsets (i_beam * per-beam size) exceed
+    // int32 for large blocks with several beams.
+    const auto i_beam = static_cast<int64_t>(blockIdx.z);
     if (t >= nsamps_out || i_dm >= ndms || i_beam >= nbeams) {
         return;
     }
